requirements.txt
djangorestframework
django-cors-headers
django-allauth
psycopg2
django-storages[azure]
python-dotenv